application/requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
gmpy2==2.1.2
idna==3.4
ipykernel==6.25.2
ipython==8.13.0
ipywidgets==8.0.4
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.3.1
jupyter-console==6.6.3
jupyter_core==5.3.2
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.1
nvidia-pyindex==1.0.9
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pillow-avif-plugin==1.4.1
pip==23.3.1
platformdirs==3.11.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==23.2.0
pyparsing==3.0.1
PyQt5
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.2
setuptools==68.0.0
sip==6.7.12
six==1.16.0
sniffio==1.2.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
widgetsnbextension==4.0.5
win-inet-pton==1.1.0
